Edits per comments from CRAN
In the Description field of the DESCRIPTION file:
Representative(SDR)
-->
Representative (SDR)

Please add small executable examples in your Rd-files to illustrate the
use of the exported functions but also enable automatic testing.
When creating the examples please keep in mind that the structure
would be desirable:
\examples{
    examples for users and checks:
    executable in < 5 sec
    \dontshow{
        examples for checks:
        executable in < 5 sec together with the examples above
        not shown to users
    }
    \donttest{
        further examples for users; not used for checks
        (f.i. data loading examples, or time consuming functions )
    }
    \dontrun{
     not used by checks, not used by example()
     adds the comment ("# Not run:") as a warning for the user.
     cannot run (e.g. because of missing additional software,
     missing API keys, ...)
    }
    if(interactive()){
        functions not intended for use in scripts,
        or are supposed to only run interactively
        (f.i. shiny)
    }
}
Your users will appreciate to see how the function is supposed to be
called for tiny examples.

Please always write TRUE and FALSE instead of T and F.
Also never name your variables T or F.